@@ -10,14 +10,14 @@
     "axios": "^0.18.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-react-native-stage-0": "^1.0.1",
-    "expo": "^23.0.0",
+    "expo": "^25.0.0",
     "immutability-helper": "^2.2.2",
     "mobx": "^3.1.16",
     "mobx-logger": "^0.6.0",
     "mobx-react": "^4.1.8",
     "native-base": "^2.2.1",
-    "react": "16.0.0-alpha.12",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-23.0.0.tar.gz",
+    "react": "16.2.0",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-25.0.0.tar.gz",
     "react-native-open-maps": "^0.2.2",
     "react-native-push-notification": "^3.0.2",
     "react-native-simple-store": "^1.3.0",
